refactor: Organizar vistas de Login y Monitor en subdirectorios
Continuando con la tarea de organización del proyecto, este commit mueve las vistas de Login y Monitor a sus propios subdirectorios dedicados.

**Cambios Realizados:**

1.  **Movimiento de Archivos:**
    - `views/2fa_verify_view.php` -> `views/login/2fa_verify_view.php`
    - `views/login_view.php` -> `views/login/login_view.php`
    - `views/monitor_view.php` -> `views/monitor/monitor_view.php`

2.  **Actualización de Referencias:**
    - Se ha actualizado `controllers/login_controller.php` para que apunte a las nuevas rutas dentro de `views/login/`.
    - Se ha actualizado `controllers/monitor_controller.php` para que apunte a la nueva ruta dentro de `views/monitor/`.

Esta refactorización mejora la estructura y mantenibilidad del directorio de vistas.
