+++ views/monitor/monitor_view.php
